Fixing footer, renaming about, adding text to hello component
