+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Hello from './components/Hello'
-import About from "./components/About";
+import Resume from "./components/Resume";
 import Projects from "./components/Projects";
 import { Switch, Route } from "react-router-dom";
 
@@ -12,7 +12,7 @@ const App = () => {
       <Header />
       <Switch>
         <Route exact path="/" component={Hello}></Route>
-        <Route exact path="/resume" component={About}></Route>
+        <Route exact path="/resume" component={Resume}></Route>
         <Route exact path="/projects" component={Projects}></Route>
         <Hello />
       </Switch>
