+++ src/components/Footer.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Container } from "semantic-ui-react";
 
 const Footer = () => {
   return (
